TradesProducer/requirements.txt
websocket-client
kafka-python
avro-python3